Create parser function to json array
This patch move the struct mydevice to cloud header so the parser don't
need to know what elements will be on queue.
